Split testing files
